2lab/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==1.10.18
passlib[bcrypt]==1.7.4
python-jose==3.3.0
cryptography==43.0.3
alembic==1.13.3
python-dotenv==1.0.1
pillow==10.4.0
numpy==2.1.2